+++ CHANGES.rst
@@ -1,14 +1,25 @@
 CHANGELOG
 =========
 
-2.1.0
+unreleased
 ----------
 
 Feature
 +++++++
 
 - add `user` option to setup and tear down mysql process as non-privileged
-  user (without root permission).
+
+Misc
+++++
+
+- Add Python 3.10 to CI
+
+2.1.0
+----------
+
+Feature
++++++++
+
 - `mysql_noproc` fixture to connect to already running mysql server
 - raise more meaningful error when the test database already exists